+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+* Added example for calling the inference endpoint with a minimal client
+* Added missing doc generation for inference examples
+
 v1.10.0 (2025-04-17)
 -------------------